// api/generated/models/Staff.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Schema for creating a new staff member.
 */
export type Staff = {
    school_id: string;
    first_name: string;
    middle_name: string;
    last_name: string;
    date_of_birth: string;
    display_name?: (string | null);
    gender?: (string | null);
    employee_number?: (string | null);
    joining_date: string;
    exit_date?: (string | null);
    primary_phone?: (string | null);
    primary_email?: (string | null);
    emergency_name?: (string | null);
    emergency_relation?: (string | null);
    emergency_phone?: (string | null);
    emergency_email?: (string | null);
    address_line1?: (string | null);
    address_line2?: (string | null);
    address_city?: (string | null);
    address_state?: (string | null);
    address_postal?: (string | null);
    address_country?: (string | null);
    blood_group?: (string | null);
    probation_end_date?: (string | null);
    muniverse_id?: (string | null);
};